+++ oauth2_http/java/com/google/auth/oauth2/ImpersonatedCredentials.java
@@ -104,9 +104,6 @@ public class ImpersonatedCredentials extends GoogleCredentials
   private static final int DEFAULT_LIFETIME_IN_SECONDS = 3600;
   private static final String CLOUD_PLATFORM_SCOPE =
       "https://www.googleapis.com/auth/cloud-platform";
-  private static final String IAM_ACCESS_TOKEN_ENDPOINT =
-      "https://iamcredentials.googleapis.com/v1/projects/-/serviceAccounts/%s:generateAccessToken";
-
   private GoogleCredentials sourceCredentials;
   private String targetPrincipal;
   private List<String> delegates;
@@ -431,14 +428,7 @@ public class ImpersonatedCredentials extends GoogleCredentials
 
   @Override
   public GoogleCredentials createScoped(Collection<String> scopes) {
-    return toBuilder()
-        .setScopes(new ArrayList<>(scopes))
-        .setLifetime(this.lifetime)
-        .setDelegates(this.delegates)
-        .setHttpTransportFactory(this.transportFactory)
-        .setQuotaProjectId(this.quotaProjectId)
-        .setIamEndpointOverride(this.iamEndpointOverride)
-        .build();
+    return toBuilder().setScopes(new ArrayList<>(scopes)).setAccessToken(null).build();
   }
 
   @Override
@@ -465,7 +455,7 @@ public class ImpersonatedCredentials extends GoogleCredentials
         .build();
   }
 
-  private ImpersonatedCredentials(Builder builder) {
+  private ImpersonatedCredentials(Builder builder) throws IOException {
     super(builder);
     this.sourceCredentials = builder.getSourceCredentials();
     this.targetPrincipal = builder.getTargetPrincipal();
@@ -480,7 +470,7 @@ public class ImpersonatedCredentials extends GoogleCredentials
     this.transportFactoryClassName = this.transportFactory.getClass().getName();
     this.calendar = builder.getCalendar();
     if (this.delegates == null) {
-      this.delegates = new ArrayList<String>();
+      this.delegates = new ArrayList<>();
     }
     if (this.scopes == null) {
       throw new IllegalStateException("Scopes cannot be null");
@@ -488,20 +478,28 @@ public class ImpersonatedCredentials extends GoogleCredentials
     if (this.lifetime > TWELVE_HOURS_IN_SECONDS) {
       throw new IllegalStateException("lifetime must be less than or equal to 43200");
     }
+
+    // Do not expect explicit universe domain, throw exception if the explicit universe domain
+    // does not match the source credential.
+    // Do nothing if it matches the source credential
+    if (isExplicitUniverseDomain()
+        && !this.sourceCredentials.getUniverseDomain().equals(builder.getUniverseDomain())) {
+      throw new IllegalStateException(
+          String.format(
+              "Universe domain %s in source credentials "
+                  + "does not match %s universe domain set for impersonated credentials.",
+              this.sourceCredentials.getUniverseDomain(), builder.getUniverseDomain()));
+    }
   }
 
+  /**
+   * Gets the universe domain for the credential.
+   *
+   * @return the universe domain from source credentials
+   */
   @Override
   public String getUniverseDomain() throws IOException {
-    try {
-      if (isExplicitUniverseDomain()) {
-        return super.getUniverseDomain();
-      }
-      return this.sourceCredentials.getUniverseDomain();
-    } catch (IOException e) {
-      // Throwing an IOException would be a breaking change, so wrap it here.
-      // This should not happen for this credential type.
-      throw new IllegalStateException(e);
-    }
+    return this.sourceCredentials.getUniverseDomain();
   }
 
   @Override
@@ -521,7 +519,14 @@ public class ImpersonatedCredentials extends GoogleCredentials
       this.sourceCredentials =
           this.sourceCredentials.createScoped(Arrays.asList(CLOUD_PLATFORM_SCOPE));
     }
-    if (isDefaultUniverseDomain()) {
+
+    // skip for SA with SSJ flow because it uses self-signed JWT
+    // and will get refreshed at initialize request step
+    // run for other source credential types or SA with GDU assert flow
+    if (!(this.sourceCredentials instanceof ServiceAccountCredentials)
+        || (isDefaultUniverseDomain()
+            && ((ServiceAccountCredentials) this.sourceCredentials)
+                .shouldUseAssertionFlowForGdu())) {
       try {
         this.sourceCredentials.refreshIfExpired();
       } catch (IOException e) {
@@ -538,7 +543,11 @@ public class ImpersonatedCredentials extends GoogleCredentials
     String endpointUrl =
         this.iamEndpointOverride != null
             ? this.iamEndpointOverride
-            : String.format(IAM_ACCESS_TOKEN_ENDPOINT, this.targetPrincipal);
+            : String.format(
+                OAuth2Utils.IAM_ACCESS_TOKEN_ENDPOINT_FORMAT,
+                getUniverseDomain(),
+                this.targetPrincipal);
+
     GenericUrl url = new GenericUrl(endpointUrl);
 
     Map<String, Object> body =
@@ -638,6 +647,9 @@ public class ImpersonatedCredentials extends GoogleCredentials
     if (!(obj instanceof ImpersonatedCredentials)) {
       return false;
     }
+    if (!super.equals(obj)) {
+      return false;
+    }
     ImpersonatedCredentials other = (ImpersonatedCredentials) obj;
     return Objects.equals(this.sourceCredentials, other.sourceCredentials)
         && Objects.equals(this.targetPrincipal, other.targetPrincipal)
@@ -651,7 +663,7 @@ public class ImpersonatedCredentials extends GoogleCredentials
 
   @Override
   public Builder toBuilder() {
-    return new Builder(this.sourceCredentials, this.targetPrincipal);
+    return new Builder(this);
   }
 
   public static Builder newBuilder() {
@@ -671,9 +683,27 @@ public class ImpersonatedCredentials extends GoogleCredentials
 
     protected Builder() {}
 
+    /**
+     * @param sourceCredentials The source credentials to use for impersonation.
+     * @param targetPrincipal The service account to impersonate.
+     * @deprecated Use {@link #Builder(ImpersonatedCredentials)} instead. This constructor will be
+     *     removed in a future release.
+     */
+    @Deprecated
     protected Builder(GoogleCredentials sourceCredentials, String targetPrincipal) {
       this.sourceCredentials = sourceCredentials;
       this.targetPrincipal = targetPrincipal;
+    }
+
+    protected Builder(ImpersonatedCredentials credentials) {
+      super(credentials);
+      this.sourceCredentials = credentials.sourceCredentials;
+      this.targetPrincipal = credentials.targetPrincipal;
+      this.delegates = credentials.delegates;
+      this.scopes = credentials.scopes;
+      this.lifetime = credentials.lifetime;
+      this.transportFactory = credentials.transportFactory;
+      this.iamEndpointOverride = credentials.iamEndpointOverride;
     }
 
     @CanIgnoreReturnValue
@@ -767,7 +797,13 @@ public class ImpersonatedCredentials extends GoogleCredentials
 
     @Override
     public ImpersonatedCredentials build() {
-      return new ImpersonatedCredentials(this);
+      try {
+        return new ImpersonatedCredentials(this);
+      } catch (IOException e) {
+        // throwing exception would be breaking change. catching instead.
+        // this should never happen.
+        throw new IllegalStateException(e);
+      }
     }
   }